Adding second set of tests
