+++ openquake/hazardlib/tests/gsim/data/ITA19/BindiEtAl2011BackboneLOW.csv
@@ -1,0 +1,10 @@
+rup_mag,rup_rake,dist_rjb,site_vs30,result_type,damping,pga
+5,-90,1,800,MEAN,5,0.053278231
+6,-90,1,800,MEAN,5,0.117944975
+7,-90,1,800,MEAN,5,0.180967402
+5,-90,10,800,MEAN,5,0.028254617
+6,-90,10,800,MEAN,5,0.071572592
+7,-90,10,800,MEAN,5,0.125659502
+5,-90,100,800,MEAN,5,0.000631889
+6,-90,100,800,MEAN,5,0.003574364
+7,-90,100,800,MEAN,5,0.014013541
