+++ openquake/hazardlib/tests/gsim/data/ITA19/BindiEtAl2011BackboneUPP.csv
@@ -1,0 +1,10 @@
+rup_mag,rup_rake,dist_rjb,site_vs30,result_type,damping,pga
+5,-90,1,800,MEAN,5,0.118572883
+6,-90,1,800,MEAN,5,0.249689513
+7,-90,1,800,MEAN,5,0.545835101
+5,-90,10,800,MEAN,5,0.062881806
+6,-90,10,800,MEAN,5,0.151519178
+7,-90,10,800,MEAN,5,0.379015038
+5,-90,100,800,MEAN,5,0.001406296
+6,-90,100,800,MEAN,5,0.007566929
+7,-90,100,800,MEAN,5,0.042267738
